fixed 32bit build

+++ test/sqlite_wrapper_tests_mocked.cpp
@@ -121,7 +121,8 @@ namespace
         EXPECT_CALL(*get_mock(), sqlite3_bind_blob64(stmt, index, NotNull(), value.size(), IsNull()))
             .WillOnce(DoAll(Invoke([value] (sqlite3_stmt*, int, const void* param_value, sqlite3_uint64 byteSize, void(*)(void*))
                                    {
-                                     const auto data{sqlite_wrapper::const_byte_span{static_cast<const std::byte*>(param_value), byteSize}};
+                                  const auto data{sqlite_wrapper::const_byte_span{static_cast<const std::byte*>(param_value),
+                                                                                  static_cast<std::size_t>(byteSize)}};
 
                                      ASSERT_THAT(data, ElementsAreArray(value));
                                    }), Return(SQLITE_OK)))
